fix a bug in proc definition; add name_proc function

+++ Code Transplant Helper/proc.h
@@ -5,5 +5,5 @@
 struct proc {
 	unsigned short code;
 	unsigned short sarg;
-	unsigned short larg;
+	unsigned int larg;
 };
